Fix warning C4018: '>=': signed/unsigned mismatch

+++ test/limits.cpp
@@ -1,4 +1,4 @@
-// Copyright 2022 Matt Borland
+// Copyright 2023 Matt Borland
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 
@@ -12,7 +12,7 @@ template <typename T>
 void test()
 {
     BOOST_TEST_GE(boost::charconv::limits<T>::max_chars10(), boost::charconv::detail::num_digits((std::numeric_limits<T>::max)()));
-    BOOST_TEST_GE(sizeof(T) * CHAR_BIT, boost::charconv::limits<T>::max_chars());
+    BOOST_TEST_GE(static_cast<int>(sizeof(T) * CHAR_BIT), boost::charconv::limits<T>::max_chars());
 }
 
 int main(void)
